Allow the cache to build in the background
A single thread worker pool is introduced, allowing workloads like
cache reloading and garbage collecting to be executed in FIFO-order.

By building the cache in the background, it allows for Jenkins
installations with a large amount of builds to serve the UI much quicker
after a restart.

On one test server with approx. 60k builds, the time from restart to UI
went from 4 minutes down to roughly 30 seconds.

While the cache is building the new method BuildCache.isCacheRefreshing()
will return true. Clients that depends on this plug-in needs to handle
this case accordingly.