Add help text for downstreamBuils step
I didn't add any config.xml since the only parameter to the step is a
RunWrapper, which is hard to visualize/provide through an input box.
So only the help text should be sufficient. Also added a note about
incomplete return value if the cache isn't fully initiated at the time
of call.